requirements.txt
kafka-python
nose
sphinx